# Optimization of Base64UrlEncode operations based (copy from Wilson)