--- a/src/client/Microsoft.Identity.Client/Internal/ClientInfo.cs
+++ b/src/client/Microsoft.Identity.Client/Internal/ClientInfo.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Identity.Client.Internal
 
             try
             {
-                return JsonHelper.DeserializeFromJson<ClientInfo>(Base64UrlHelpers.DecodeToBytes(clientInfo));
+                return JsonHelper.DeserializeFromJson<ClientInfo>(Base64UrlHelpers.DecodeBytes(clientInfo));
             }
             catch (Exception exc)
             {