--- a/src/client/Microsoft.Identity.Client/Internal/ClientInfo.cs
+++ b/src/client/Microsoft.Identity.Client/Internal/ClientInfo.cs
@@ -8,6 +8,8 @@ using Microsoft.Identity.Json;
 
 namespace Microsoft.Identity.Client.Internal
 {
+
+
     [JsonObject]
     [Preserve(AllMembers = true)]
     internal class ClientInfo