--- a/tests/Microsoft.Identity.Test.Performance/Program.cs
+++ b/tests/Microsoft.Identity.Test.Performance/Program.cs
@@ -3,6 +3,7 @@
 
 using System;
 using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Diagnosers;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Running;
 
@@ -12,14 +13,17 @@ namespace Microsoft.Identity.Test.Performance
     {
         static void Main(string[] args)
         {
-            BenchmarkRunner.Run<AcquireTokenForClientLargeCacheTests>(
+            BenchmarkRunner.Run<Base64Encode>(
                 DefaultConfig.Instance
                     .WithOptions(ConfigOptions.DontOverwriteResults)
-                    
+                    .AddDiagnoser(MemoryDiagnoser.Default)
                     .AddJob(
                         Job.Default
-                            .WithId("Job-PerfTests")                            
-                            ));
+                            .WithId("Job-PerfTests")
+                    //        .WithGcMode(new GcMode() { Server = true })
+
+));
+
 
             Console.ReadKey();
         }