Fixes # .
<!-- Add the issue number above. If this PR only partially fixes the issue, remove the Fixes word above so that the issue is not automatically closed. -->

**Changes proposed in this request**
Copy UrlEncoders from Wilson
Add     `<LangVersion>latest</LangVersion>` - this will allow us to use newer language features, but we are still bound to c# 7.3 due to net45 and netstandard1.3 ... Once we stop supporting those, we can move to C# 8 or 9 consistently. 

**Testing**
perf test. existing ut cover functionality

**Performance impact**
Encode operations are 2x+ faster and consume 1/2 memory. Decode operations are more or less the same.


|              Method |           Job |       Runtime | Server |    StdDev |     Median | Allocated |
|-------------------- |-------------- |-------------- |------- |----------:|-----------:|----------:|
| Encoding_String_Old |    .NET 4.7.2 |    .NET 4.7.2 |  False |  10.17 ns | 1,080.9 ns |    2006 B |
|   Encode_String_New |    .NET 4.7.2 |    .NET 4.7.2 |  False |   7.83 ns |   487.5 ns |    1107 B |
|     Encode_Byte_Old |    .NET 4.7.2 |    .NET 4.7.2 |  False |  71.60 ns |   970.7 ns |    1789 B |
|     Encode_Byte_New |    .NET 4.7.2 |    .NET 4.7.2 |  False |   8.97 ns |   350.5 ns |     891 B |
|    Decode_Bytes_Old |    .NET 4.7.2 |    .NET 4.7.2 |  False |   8.72 ns |   851.9 ns |     530 B |
|    Decode_Bytes_New |    .NET 4.7.2 |    .NET 4.7.2 |  False |   9.58 ns |   782.6 ns |     530 B |
|   Decode_String_Old |    .NET 4.7.2 |    .NET 4.7.2 |  False |  86.36 ns | 1,135.5 ns |     883 B |
|   Decode_String_New |    .NET 4.7.2 |    .NET 4.7.2 |  False |  12.93 ns | 1,054.3 ns |     883 B |
| Encoding_String_Old | .NET Core 3.1 | .NET Core 3.1 |  False |  24.56 ns |   985.7 ns |    1072 B |
|   Encode_String_New | .NET Core 3.1 | .NET Core 3.1 |  False |  17.43 ns |   492.4 ns |    1096 B |
|     Encode_Byte_Old | .NET Core 3.1 | .NET Core 3.1 |  False | 376.60 ns | 1,114.4 ns |     856 B |
|     Encode_Byte_New | .NET Core 3.1 | .NET Core 3.1 |  False |  52.15 ns |   450.8 ns |     880 B |
|    Decode_Bytes_Old | .NET Core 3.1 | .NET Core 3.1 |  False |  11.78 ns |   739.5 ns |     512 B |
|    Decode_Bytes_New | .NET Core 3.1 | .NET Core 3.1 |  False | 152.94 ns |   697.7 ns |     512 B |
|   Decode_String_Old | .NET Core 3.1 | .NET Core 3.1 |  False | 175.25 ns |   996.4 ns |     856 B |
|   Decode_String_New | .NET Core 3.1 | .NET Core 3.1 |  False | 153.59 ns |   894.4 ns |     856 B |
| Encoding_String_Old | Job-PerfTests | .NET Core 3.1 |   True |  44.78 ns | 1,012.5 ns |    1072 B |
|   Encode_String_New | Job-PerfTests | .NET Core 3.1 |   True |  13.91 ns |   526.5 ns |    1096 B |
|     Encode_Byte_Old | Job-PerfTests | .NET Core 3.1 |   True |   9.36 ns |   736.1 ns |     856 B |
|     Encode_Byte_New | Job-PerfTests | .NET Core 3.1 |   True |   6.22 ns |   370.5 ns |     880 B |
|    Decode_Bytes_Old | Job-PerfTests | .NET Core 3.1 |   True |  16.89 ns |   743.9 ns |     512 B |
|    Decode_Bytes_New | Job-PerfTests | .NET Core 3.1 |   True |   6.71 ns |   628.6 ns |     512 B |
|   Decode_String_Old | Job-PerfTests | .NET Core 3.1 |   True |  17.95 ns |   993.5 ns |     856 B |
|   Decode_String_New | Job-PerfTests | .NET Core 3.1 |   True |  17.39 ns |   932.0 ns |     856 B |